Experiment: Create useTimeseries hook.
