+++ src/App.scss
@@ -4507,6 +4507,7 @@ footer {
   font-family: 'archia';
   color: $gray-mid;
   overflow-y: scroll;
+  height: 10vh;
 
   .date {
     cursor: pointer;
